Remove gitlab ci retry
There is no way in gitlab to determine from the stage environment the
number of the current retry attempt. This presents a problem with the
kitmaker webhook callback where a timeout on the internal cuda-repo
(which happens often) causes the webhook to be called with a failed
status, but the gitlab ci second attempt (untracked by kitmaker)
succeeds.

Ugh! A request to add this to gitlab ci was closed here:
https://gitlab.com/gitlab-org/gitlab/-/issues/195618